separted search form from search results
